+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added to `vtex.formatted-price@0.x`.
+- New variable `sellingPriceValueFormatted` in the Selling Price Renderer for ICU messages, aiding in the creation of custom messages.
+
 ## [1.27.0] - 2022-03-08
 
 ### Added